progress float64

+++ client/client.go
@@ -138,7 +138,7 @@ type Torrent struct {
 	NumLeechs              int64   `json:"num_leechs"`
 	NumSeeds               int64   `json:"num_seeds"`
 	Priority               int64   `json:"priority"`
-	Progress               int64   `json:"progress"`
+	Progress               float64 `json:"progress"`
 	Ratio                  float64 `json:"ratio"`
 	RatioLimit             int64   `json:"ratio_limit"`
 	SavePath               string  `json:"save_path"`
